Extract streams lib and add unit tests.
